image post processing
This code post-processes a three-dimensional video. Videos created by FIXIE have 2 channels for the readouts of the two cameras, making 2 frames per timepoint and adding a blank image in red when no laser is firing.

This code sorts out the FIXIE image data and the end result is a 3 color image (with Z information) and a MaxZ projection.